requirements.txt
Flask_RESTful==0.3.8
requests==2.24.0
Flask==1.1.2
pandas==1.1.3
pyspark==3.0.1
python-dotenv==0.15.0
gunicorn
